refactor: codebase restructured to be more modular + middlewares converted to a tower based service & layer
