+++ core/server/src/middleware/mod.rs
@@ -1,0 +1,3 @@
+pub mod one_time_payment;
+pub mod payment_channel;
+pub mod stream_payment;
