+++ core/server/src/middleware/payment_channel/verify.rs
@@ -2,11 +2,8 @@
 use std::time::{SystemTime, UNIX_EPOCH};
 
 use alloy::{
-    dyn_abi::DynSolType,
     hex::{self},
-    primitives::{Address, Signed, U256},
-    providers::{Provider, ProviderBuilder},
-    sol,
+    primitives::U256,
     transports::http::reqwest::Url,
 };
 
@@ -14,19 +11,18 @@ use alloy::{
 use js_sys::Date;
 
 use crate::{
-    channel::ChannelState,
     error::AuthError,
-    types::{
-        tx::{SignedStream, StreamsConfig},
-        OneTimePaymentConfig, PaymentChannel, SignedPaymentTx, SignedRequest,
+    middleware::payment_channel::{
+        channel::ChannelState,
+        types::{PaymentChannel, SignedRequest},
+        utils::create_channel_message,
     },
-    utils::{create_channel_message, create_tx_message, helpers::create_stream_message},
 };
 
 pub async fn verify_and_update_channel(
     state: &ChannelState,
     mut request: SignedRequest,
-) -> Result<PaymentChannel, AuthError> {
+) -> Result<(PaymentChannel, bool), AuthError> {
     println!("\n=== verify_and_update_channel ===");
     println!("Payment amount: {}", request.payment_amount);
     println!(
@@ -142,7 +138,7 @@ pub async fn verify_and_update_channel(
     );
 
     println!("API request authorized");
-    Ok(request.payment_channel.clone())
+    Ok((request.payment_channel.clone(), true))
 }
 
 // Verify the channel and return the updated channel
@@ -151,7 +147,7 @@ pub async fn verify_channel(
     rpc_url: Url,
     mut request: SignedRequest,
     current_channel: Option<PaymentChannel>,
-) -> Result<PaymentChannel, AuthError> {
+) -> Result<(PaymentChannel, bool), AuthError> {
     println!("\n=== verify_channel ===");
     println!("Payment amount: {}", request.payment_amount);
     println!(
@@ -259,167 +255,5 @@ pub async fn verify_channel(
     request.payment_channel.balance -= request.payment_amount;
 
     println!("API request authorized");
-    Ok(request.payment_channel.clone())
-}
-
-// For one time payment verification
-pub async fn verify_tx(
-    signed_tx: SignedPaymentTx,
-    config: OneTimePaymentConfig,
-) -> Result<bool, AuthError> {
-    // creating the message
-    let reconstructed_message = create_tx_message(signed_tx.tx_hash);
-    println!("Message: 0x{}", hex::encode(&reconstructed_message));
-
-    let signature = signed_tx.signature;
-    println!("Signature: 0x{}", hex::encode(&signature.as_bytes()));
-
-    // recovering the address from the signature
-    let recovered = match signature.recover_address_from_msg(reconstructed_message) {
-        Ok(address) => address,
-        Err(_) => return Err(AuthError::InvalidSignature),
-    };
-    println!("Recovered address: {}", recovered);
-
-    let provider = ProviderBuilder::new().on_http(config.rpc_url.parse().unwrap());
-
-    // Fetching the info for transaction
-    let tx_receipt = match provider
-        .get_transaction_receipt(signed_tx.tx_hash)
-        .await
-        .map_err(|e| AuthError::ContractError(e.to_string()))?
-    {
-        Some(tx_receipt) => tx_receipt,
-        None => {
-            println!("Failed: Transaction not found");
-            return Err(AuthError::TransactionNotFound);
-        }
-    };
-
-    // Verifying recovered address against the sender for the transaction
-    if recovered != tx_receipt.from {
-        println!("Failed: Recovered address mismatch");
-        return Err(AuthError::InvalidSignature);
-    }
-
-    // Match the contract interacted with to be the token contract
-    match tx_receipt.to {
-        Some(to) => {
-            if to != config.token_address {
-                return Err(AuthError::InvalidTransaction(
-                    "Invalid token contract address".to_string(),
-                ));
-            }
-        }
-        None => {
-            println!("Failed: To address not found");
-            return Err(AuthError::InvalidTransaction(
-                "To address not found".to_string(),
-            ));
-        }
-    }
-
-    let receipt = match tx_receipt.inner.as_receipt() {
-        Some(receipt) => receipt,
-        None => {
-            return Err(AuthError::InvalidTransaction(
-                "Receipt not found".to_string(),
-            ))
-        }
-    };
-
-    let transfer_log = match receipt.logs.first() {
-        Some(log) => log,
-        None => return Err(AuthError::InvalidTransaction("Log not found".to_string())),
-    };
-
-    // Check if the log is a transfer log & verify the topics and data
-    if transfer_log.address() == config.token_address {
-        match transfer_log.topics().get(2) {
-            Some(t) => {
-                let to = Address::from_word(t.clone());
-                let data = &transfer_log.data().data;
-                let data_type = DynSolType::Uint(256);
-                let decoded = data_type
-                    .abi_decode(&data)
-                    .map_err(|e| AuthError::ContractError(e.to_string()))?;
-
-                let (amount, _) = match decoded.as_uint() {
-                    Some(amount) => amount,
-                    None => {
-                        return Err(AuthError::InvalidTransaction(
-                            "Amount couldn't be parsed from event".to_string(),
-                        ))
-                    }
-                };
-
-                if to != config.recipient || amount != config.amount {
-                    return Err(AuthError::InvalidTransaction(
-                        "Invalid recipient or amount".to_string(),
-                    ));
-                }
-            }
-            None => return Err(AuthError::InvalidTransaction("Topic not found".to_string())),
-        }
-    }
-
-    Ok(true)
-}
-
-sol!(
-    #[allow(missing_docs)]
-    #[sol(rpc)]
-    CFAv1Forwarder,
-    "src/abi/CFAv1Forwarder.json"
-);
-
-pub async fn verify_stream(stream: SignedStream, config: StreamsConfig) -> Result<bool, AuthError> {
-    // Creating the message
-    let reconstructed_message = create_stream_message(stream.sender);
-    println!("Message: 0x{}", hex::encode(&reconstructed_message));
-
-    let signature = stream.signature;
-    println!("Signature: 0x{}", hex::encode(&signature.as_bytes()));
-
-    // Recovering the address from the signature
-    let recovered = match signature.recover_address_from_msg(reconstructed_message) {
-        Ok(address) => address,
-        Err(_) => return Err(AuthError::InvalidSignature),
-    };
-    println!("Recovered address: {}", recovered);
-
-    // Verify the recovered address against the sender for the stream
-    if recovered != stream.sender {
-        println!("Failed: Recovered address mismatch");
-        return Err(AuthError::InvalidSignature);
-    }
-
-    let provider = ProviderBuilder::new().on_http(config.rpc_url.parse().unwrap());
-
-    let cfav1_forwarder = CFAv1Forwarder::new(config.cfa_forwarder, provider);
-
-    // Fetch the stream flow from sender to recipient, if it exists, using CFAv1Forwarder
-    let flow_info = cfav1_forwarder
-        .getFlowInfo(config.token_address, stream.sender, config.recipient)
-        .call()
-        .await
-        .map_err(|e| AuthError::ContractError(e.to_string()))?;
-
-    // Check if the flow exists
-    if flow_info.flowrate == Signed::ZERO {
-        println!("Failed: No stream flow found");
-        return Err(AuthError::InvalidStream("No stream flow found".to_string()));
-    } else {
-        println!("Stream flow found");
-        println!("Flow rate: {}", flow_info.flowrate);
-        // check the flowRate matches with what recipient expects
-        if flow_info.flowrate != config.amount {
-            println!("Failed: Invalid stream flow rate");
-            return Err(AuthError::InvalidStream(
-                "Invalid stream flow rate".to_string(),
-            ));
-        }
-    }
-
-    Ok(true)
+    Ok((request.payment_channel.clone(), true))
 }
