--- core/server/src/types/mod.rs
@@ -1,7 +1,0 @@
-pub mod channel;
-pub mod tx;
-
-pub use alloy::primitives::U256;
-pub use alloy::transports::http::reqwest::Url;
-pub use channel::{PaymentChannel, SignedRequest};
-pub use tx::{OneTimePaymentConfig, SignedPaymentTx};
