+++ core/server/src/utils/mod.rs
@@ -1,5 +1,1 @@
-pub mod headers;
-pub mod helpers;
 
-pub use headers::{modify_headers_axum, parse_headers};
-pub use helpers::{create_channel_message, create_tx_message};
